backend/requirements.txt
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4

pytest==8.3.4
pytest-django==4.9.0
pytest-html-reporter==0.2.9
selenium==4.27.1
